Group Project/ghostapi/requirements.txt
amqp==1.4.8
anyjson==0.3.3
billiard==3.3.0.22
celery==3.1.19
csvImporter==0.1.3.2
cycler==0.9.0
dj-database-url==0.3.0
Django==1.8.5
django-celery==3.1.17
django-cors-headers==1.1.0
djangorestframework==3.2.4
kombu==3.0.32
matplotlib==1.5.0
mpld3==0.2
numpy==1.10.1
pandas==0.17.0
pika==0.10.0
postcode.io==0.1.1
pyparsing==2.0.7
python-dateutil==2.4.2
pytz==2015.7
requests==2.8.1
six==1.10.0
wheel==0.24.0
